config/requirements-ai-tools.txt
# AI Tools Dependencies
# This file contains Python dependencies for the AI tools integration

# Core
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0

# Graph Databases
neo4j>=5.12.0
redis>=5.0.1

# AI/ML
openai>=1.3.0
langchain>=0.0.340
llama-index>=0.9.15

# CrewAI
crewai>=0.1.0

# Utilities
tqdm>=4.66.1
loguru>=0.7.2
pydantic>=2.4.2

# Web
fastapi>=0.104.1
uvicorn>=0.24.0

# Data Processing
numpy>=1.26.0
pandas>=2.1.1

# Testing
pytest>=7.4.2
pytest-cov>=4.1.0
